rebased on latest collectives refactor move barrier to init
